// services/ssi-issuance/src/infrastructure/modules/acapy/present-proof-v2/types/pres-exchange-record-state.enum.ts
/****************************************************************************
 * Copyright 2022 Spherity GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ***************************************************************************/

export enum PresentationExchangeRecordState {
  PROPOSAL_SENT = 'proposal-sent',
  PROPOSAL_RECEIVED = 'proposal-received',
  REQUEST_SENT = 'request-sent',
  REQUEST_RECEIVED = 'request-received',
  PRESENTATION_SENT = 'presentation-sent',
  PRESENTATION_RECEIVED = 'presentation-received',
  DONE = 'done',
  ABANDONED = 'abandoned',
}